added matrix rain animation and music working

+++ src/11_Ali/src/memory/malloc.c
@@ -14,7 +14,7 @@ uint32_t memory_used = 0;
 // Initialize the kernel memory manager
 void init_kernel_memory(uint32_t* kernel_end)
 {
-    last_alloc = kernel_end + 0x1000;
+    last_alloc = (uint32_t)kernel_end + 0x1000;
     heap_begin = last_alloc;
     pheap_end = 0x400000;
     pheap_begin = pheap_end - (MAX_PAGE_ALIGNED_ALLOCS * 4096);
@@ -46,7 +46,7 @@ void free(void *mem)
 // Free a block of page-aligned memory
 void pfree(void *mem)
 {
-    if(mem < pheap_begin || mem > pheap_end) return;
+    if((uint32_t)mem < pheap_begin || (uint32_t)mem > pheap_end) return;
 
     // Determine the page ID
     uint32_t ad = (uint32_t)mem;
